[BACKLOG-18999] Refactoring ActionResources into a 'true' resource class (REST path alone) and a services class that holds the execution logic
